chore: move high level classes to common workspace
Move DesktopManager, DesktopConnection, ExtensionPairing, and DesktopController to common workspace.
