+++ packages/app/src/app/globals.ts
@@ -1,11 +1,13 @@
 import 'global-agent/bootstrap';
 import '../browser/browser-init';
+import './controller-init';
 import { webcrypto } from 'node:crypto';
 import { Headers } from 'node-fetch';
 import setupSentry from '../../submodules/extension/app/scripts/lib/setupSentry';
 import { getDesktopVersion } from '../utils/version';
 
 declare const global: typeof globalThis & {
+  isDesktopApp: boolean;
   stateHooks: Record<string, any>;
   sentry: unknown;
 };
@@ -13,6 +15,7 @@ declare const global: typeof globalThis & {
 if (!global.self) {
   global.self = {} as unknown as Window & typeof globalThis;
   global.crypto = webcrypto as any;
+  global.isDesktopApp = true;
 
   global.navigator = {
     userAgent: 'Firefox',
