+++ packages/app/src/shared/pairing.ts
@@ -1,159 +1,17 @@
 import { Duplex } from 'stream';
-///: BEGIN:ONLY_INCLUDE_IN(desktopapp)
 import EventEmitter from 'events';
-///: END:ONLY_INCLUDE_IN
 import log from 'loglevel';
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore
 import ObjectMultiplex from 'obj-multiplex';
-import { browser } from '@metamask/desktop/dist/browser';
-import { MESSAGE_ACKNOWLEDGE } from '@metamask/desktop/dist/constants';
-import {
-  ///: BEGIN:ONLY_INCLUDE_IN(desktopapp)
-  acknowledge,
-  ///: END:ONLY_INCLUDE_IN
-  waitForAcknowledge,
-  waitForMessage,
-} from '@metamask/desktop/dist/utils/stream';
-import TOTP from '@metamask/desktop/dist/utils/totp';
-import { createKey } from '@metamask/desktop/dist/encryption/symmetric';
-import { hashString } from '@metamask/desktop/dist/utils/crypto';
+import { acknowledge } from '@metamask/desktop/dist/utils/stream';
 import {
   PairingKeyRequestMessage,
   PairingKeyResponseMessage,
-  PairingRequestMessage,
   PairingResultMessage,
-  PairingKeyStatus,
 } from '@metamask/desktop/dist/types';
 import * as rawState from '@metamask/desktop/dist/utils/state';
 
-const createStreams = (stream: Duplex) => {
-  const multiplex = new ObjectMultiplex();
-  const requestStream = multiplex.createStream('request');
-  const keyStream = multiplex.createStream('key');
-
-  stream.pipe(multiplex).pipe(stream);
-
-  return { requestStream, keyStream };
-};
-
-///: BEGIN:ONLY_INCLUDE_IN(desktopextension)
-export class ExtensionPairing {
-  private requestStream: Duplex;
-
-  private keyStream: Duplex;
-
-  private transferState: () => Promise<void>;
-
-  constructor(stream: Duplex, transferState: () => Promise<void>) {
-    const streams = createStreams(stream);
-
-    this.requestStream = streams.requestStream;
-    this.keyStream = streams.keyStream;
-    this.transferState = transferState;
-  }
-
-  public static generateOTP(): string {
-    const value = TOTP.generate() as string;
-    log.debug('Generated OTP', value);
-    return value;
-  }
-
-  public init() {
-    this.requestStream.on('data', (data: PairingRequestMessage | string) => {
-      if (data === MESSAGE_ACKNOWLEDGE) {
-        return;
-      }
-
-      this.onRequestMessage(data as PairingRequestMessage);
-    });
-
-    return this;
-  }
-
-  public async checkPairingKeyMatch(): Promise<PairingKeyStatus> {
-    log.debug('Validating pairing key');
-
-    const requestPairingKey: PairingKeyRequestMessage = {
-      isRequestPairingKey: true,
-    };
-
-    this.keyStream.write(requestPairingKey);
-
-    // Wait for desktop pairing key
-    const response = await waitForMessage<PairingKeyResponseMessage>(
-      this.keyStream,
-    );
-
-    const desktopPairingKey = response.pairingKey;
-
-    if (!desktopPairingKey) {
-      log.debug('Desktop has no pairing key');
-      return PairingKeyStatus.MISSING;
-    }
-
-    const desktopPairingKeyHash = await hashString(desktopPairingKey, {
-      isHex: true,
-    });
-
-    const extensionPairingKeyHash = (await rawState.getDesktopState())
-      .pairingKeyHash;
-
-    const isMatch = extensionPairingKeyHash === desktopPairingKeyHash;
-
-    log.debug('Completed pairing key check', isMatch);
-
-    if (isMatch) {
-      return PairingKeyStatus.MATCH;
-    }
-    return PairingKeyStatus.NO_MATCH;
-  }
-
-  private async onRequestMessage(pairingRequest: PairingRequestMessage) {
-    log.debug('Received pairing request message', pairingRequest);
-
-    const isValidOTP = TOTP.validate(pairingRequest.otp);
-
-    if (!isValidOTP) {
-      log.debug('Received invalid OTP');
-
-      const pairingResultMessage: PairingResultMessage = {
-        isDesktopEnabled: false,
-      };
-
-      this.requestStream.write(pairingResultMessage);
-      return;
-    }
-
-    const pairingKey = await createKey();
-    const pairingKeyHash = await hashString(pairingKey, { isHex: true });
-
-    await rawState.setDesktopState({
-      desktopEnabled: true,
-      pairingKeyHash,
-    });
-
-    const pairingResultMessage: PairingResultMessage = {
-      isDesktopEnabled: true,
-      pairingKey,
-    };
-
-    this.requestStream.write(pairingResultMessage);
-
-    await waitForAcknowledge(this.requestStream);
-
-    log.debug('Saved pairing key', { pairingKey, pairingKeyHash });
-
-    await this.transferState();
-
-    browser.runtime.reload();
-
-    log.debug('Paired with desktop');
-  }
-}
-///: END:ONLY_INCLUDE_IN
-
-///: BEGIN:ONLY_INCLUDE_IN(desktopapp)
 export class DesktopPairing extends EventEmitter {
   private requestStream: Duplex;
 
@@ -162,7 +20,7 @@ export class DesktopPairing extends EventEmitter {
   constructor(stream: Duplex) {
     super();
 
-    const streams = createStreams(stream);
+    const streams = this.createStreams(stream);
 
     this.requestStream = streams.requestStream;
     this.keyStream = streams.keyStream;
@@ -219,5 +77,14 @@ export class DesktopPairing extends EventEmitter {
 
     this.keyStream.write(response);
   }
+
+  private createStreams(stream: Duplex) {
+    const multiplex = new ObjectMultiplex();
+    const requestStream = multiplex.createStream('request');
+    const keyStream = multiplex.createStream('key');
+
+    stream.pipe(multiplex).pipe(stream);
+
+    return { requestStream, keyStream };
+  }
 }
-///: END:ONLY_INCLUDE_IN
