+++ packages/app/test/mocks.ts
@@ -1,39 +1,22 @@
 import { Duplex, EventEmitter } from 'stream';
 import ObjectMultiplex from 'obj-multiplex';
 import { WebSocketServer } from 'ws';
-import { ObservableStore } from '@metamask/obs-store';
 import ElectronStore from 'electron-store';
-import { VersionCheck } from '@metamask/desktop/dist/version-check';
 import {
-  BrowserWebSocket,
   NodeWebSocket,
   WebSocketStream,
 } from '@metamask/desktop/dist/web-socket-stream';
-import {
-  ConnectionType,
-  RemotePort,
-  TestConnectionResult,
-} from '@metamask/desktop/dist/types';
+import { ConnectionType } from '@metamask/desktop/dist/types';
 import ExtensionConnection from '../src/app/extension-connection';
-import DesktopConnection from '../src/extension/desktop-connection';
-import { DesktopPairing, ExtensionPairing } from '../src/shared/pairing';
-import NotificationManager from '../submodules/extension/app/scripts/lib/notification-manager';
-import ExtensionPlatform from '../submodules/extension/app/scripts/platforms/extension';
+import { DesktopPairing } from '../src/shared/pairing';
 
-export const PUBLIC_KEY_MOCK = 'testPublicKey';
-export const PRIVATE_KEY_MOCK = 'testPrivateKey';
 export const DECRYPTED_STRING_MOCK = 'testDecryptedData';
-export const DECRYPTED_BUFFER_MOCK = Buffer.from(DECRYPTED_STRING_MOCK);
-export const ENCRYPTED_STRING_MOCK = 'testEncryptedData';
 export const PROPERTY_MOCK = 'test';
 export const PROPERTY_2_MOCK = 'test2';
 export const VALUE_MOCK = 'value';
 export const VALUE_2_MOCK = 'value2';
 export const DATA_MOCK = { [PROPERTY_MOCK]: VALUE_MOCK };
 export const DATA_2_MOCK = { [PROPERTY_2_MOCK]: VALUE_2_MOCK };
-export const JSON_MOCK = '{"test":"value"}';
-export const STRING_DATA_MOCK = 'testStringData';
-export const STRING_DATA_BUFFER_MOCK = Buffer.from(STRING_DATA_MOCK);
 export const REMOTE_PORT_NAME_MOCK = 'testPort';
 export const REMOTE_PORT_SENDER_MOCK = { test2: 'value2' };
 export const CLIENT_ID_MOCK = 3;
@@ -42,48 +25,16 @@ export const PORT_MOCK = 123;
 export const PASSWORD_MOCK = 'testPassword';
 export const ENCRYPTED_CYPHER_MOCK = Buffer.from('encryptedCypherTest');
 export const ENCRYPTED_CYPHER_FILE_MOCK = 'encryptedCypherFileTest';
-export const STREAM_MOCK = 'testStream';
-export const TYPE_MOCK = 'testType';
-export const METHOD_MOCK = 'testMethod';
-export const RESULT_MOCK = 'testResult';
-export const KEY_MOCK = {} as CryptoKey;
-export const EXPORTED_KEY_MOCK = Buffer.from([1, 2, 3]);
-export const ENCRYPTED_BUFFER_MOCK = Buffer.from([4, 5, 6]);
-export const IV_BUFFER_MOCK = Buffer.from([7, 8, 9]);
-export const JSON_RPC_ID_MOCK = 123456;
 export const ARGS_MOCK = ['test123', 123, true];
 export const UUID_MOCK = '6328e6ae-f867-4876-af6f-22a44efbe251';
 export const OTP_MOCK = '123456';
-export const WRONG_OTP_MOCK = '654321';
 export const VERSION_MOCK = '123.456.789.012';
 export const VERSION_2_MOCK = '456.123.789.012';
-export const HASH_BUFFER_MOCK = Buffer.from([10, 11, 12]);
-export const HASH_BUFFER_2_MOCK = Buffer.from([10, 11, 13]);
-export const IV_HEX_MOCK = Buffer.from(IV_BUFFER_MOCK).toString('hex');
-export const HASH_BUFFER_HEX_MOCK = HASH_BUFFER_MOCK.toString('hex');
-export const HASH_BUFFER_2_HEX_MOCK = HASH_BUFFER_2_MOCK.toString('hex');
-
-export const EXPORTED_KEY_HEX_MOCK =
-  Buffer.from(EXPORTED_KEY_MOCK).toString('hex');
-
-export const ENCRYPTED_HEX_MOCK = Buffer.from(ENCRYPTED_BUFFER_MOCK).toString(
-  'hex',
-);
 
 export const NEW_CONNECTION_MESSAGE_MOCK = {
   clientId: CLIENT_ID_MOCK,
   connectionType: ConnectionType.INTERNAL,
   remotePort: { name: REMOTE_PORT_NAME_MOCK, sender: REMOTE_PORT_SENDER_MOCK },
-};
-
-export const TEST_CONNECTION_RESULT_MOCK: TestConnectionResult = {
-  isConnected: true,
-  versionCheck: {
-    extensionVersion: '123',
-    desktopVersion: '456',
-    isDesktopVersionValid: true,
-    isExtensionVersionValid: false,
-  },
 };
 
 export const createWebSocketServerMock = (): jest.Mocked<WebSocketServer> =>
@@ -100,14 +51,6 @@ export const createWebSocketNodeMock = (): jest.Mocked<NodeWebSocket> =>
     close: jest.fn(),
     removeListener: jest.fn(),
   } as unknown as jest.Mocked<NodeWebSocket>);
-
-export const createWebSocketBrowserMock = (): jest.Mocked<BrowserWebSocket> =>
-  ({
-    readyState: 1,
-    addEventListener: jest.fn(),
-    send: jest.fn(),
-    close: jest.fn(),
-  } as unknown as jest.Mocked<BrowserWebSocket>);
 
 export const createStreamMock = (): jest.Mocked<Duplex> =>
   ({
@@ -130,24 +73,6 @@ export const createWebSocketStreamMock = (): jest.Mocked<WebSocketStream> =>
     init: jest.fn(),
     removeListener: jest.fn(),
   } as unknown as jest.Mocked<WebSocketStream>);
-
-export const createRemotePortMock = (): jest.Mocked<RemotePort> =>
-  ({
-    name: REMOTE_PORT_NAME_MOCK,
-    sender: REMOTE_PORT_SENDER_MOCK,
-    onMessage: {
-      addListener: jest.fn(),
-    },
-    onDisconnect: {
-      addListener: jest.fn(),
-    },
-  } as unknown as jest.Mocked<RemotePort>);
-
-export const createNotificationManagerMock =
-  (): jest.Mocked<NotificationManager> =>
-    ({
-      showPopup: jest.fn(),
-    } as unknown as jest.Mocked<NotificationManager>);
 
 export const createMultiplexMock = (): jest.Mocked<ObjectMultiplex & Duplex> =>
   ({
@@ -173,40 +98,10 @@ export const createExtensionConnectionMock =
       disable: jest.fn(),
     } as any);
 
-export const createDesktopConnectionMock = (): jest.Mocked<DesktopConnection> =>
-  ({
-    transferState: jest.fn(),
-    on: jest.fn(),
-    disconnect: jest.fn(),
-    getDesktopVersion: jest.fn(),
-    createStream: jest.fn(),
-    removeAllListeners: jest.fn(),
-    checkPairingKey: jest.fn(),
-  } as any);
-
-export const createObservableStoreMock = (): jest.Mocked<ObservableStore> =>
-  ({
-    getState: jest.fn(),
-    updateState: jest.fn(),
-  } as any);
-
-export const createExtensionPairingMock = (): jest.Mocked<ExtensionPairing> =>
-  ({
-    generateOTP: jest.fn(),
-    checkPairingKeyMatch: jest.fn(),
-    init: jest.fn(),
-  } as any);
-
 export const createElectronStoreMock = (): jest.Mocked<ElectronStore> =>
   ({
     clear: jest.fn(),
   } as any);
-
-export const createExtensionVersionCheckMock = (): jest.Mocked<VersionCheck> =>
-  ({ check: jest.fn() } as any);
-
-export const createExtensionPlatformMock = (): jest.Mocked<ExtensionPlatform> =>
-  ({ getVersion: jest.fn() } as any);
 
 export const createDesktopPairingMock = (): jest.Mocked<DesktopPairing> =>
   ({ init: jest.fn() } as any);
