+++ packages/common/package.json
@@ -16,8 +16,12 @@
     "test:coverage": "jest --coverage"
   },
   "dependencies": {
+    "@metamask/obs-store": "^5.0.0",
     "eciesjs": "^0.3.15",
+    "end-of-stream": "^1.4.4",
+    "extension-port-stream": "^2.0.0",
     "loglevel": "^1.8.0",
+    "obj-multiplex": "^1.0.0",
     "otpauth": "^8.0.3",
     "uuid": "^9.0.0",
     "ws": "^7.4.6"
@@ -30,6 +34,7 @@
     "@metamask/eslint-config-nodejs": "^9.0.0",
     "@metamask/eslint-config-typescript": "^9.0.1",
     "@tsconfig/node16": "^1.0.3",
+    "@types/end-of-stream": "^1.4.1",
     "@types/jest": "^29.1.2",
     "@types/node": "^17.0.21",
     "@types/uuid": "^8.3.4",
