+++ packages/common/src/browser/browser-proxy.test.ts
@@ -12,17 +12,13 @@ import { registerResponseStream } from '.';
 
 jest.mock('loglevel');
 
-jest.mock(
-  './browser-polyfill',
-  () => ({
-    browser: {
-      browserAction: {
-        setBadgeText: jest.fn(),
-      },
+jest.mock('./browser-polyfill', () => ({
+  browser: {
+    browserAction: {
+      setBadgeText: jest.fn(),
     },
-  }),
-  { virtual: true },
-);
+  },
+}));
 
 describe('Browser Proxy', () => {
   const browserMock = browser as any;
