+++ packages/common/src/controllers/init.ts
@@ -1,0 +1,10 @@
+import desktopManager from '../desktop-manager';
+import { Pairing } from '../pairing';
+
+const generateOTP = () => Pairing.generateOTP();
+const testConnection = () => desktopManager.testConnection();
+
+export const extensionLogic = {
+  generateOTP,
+  testConnection,
+};
